mosquitto: force custom SOVERSION (avoid clashes with 3rd party libmosquitto)
Previously this recipe provided libmosquitto.so.1. In some cases the
same SONAME may be provided independently by a 3rd party recipe (e.g.
a prebuilt WiFi driver etc which bundles its own prebuilt version of
libmosquitto).

To avoid any chance of accidentally running an application built with
this version of libmosquitto with a 3rd party libmosquitto library
(or vice versa) change the SONAME for this libmosquitto.so. This will
allow both versions of libmosquitto to be installed in parallel, if
required.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>